+++ src/pdf_structures/object_config.cpp
@@ -6,7 +6,7 @@
 namespace sru::pdf {
 int DATE_PROVIDER = -1;
 std::vector<ObjectConfig> ObjectConfigPool;
-auto getObjectConfig(int id) -> std::optional<ObjectConfig> {
+auto getObjectConfig(size_t id) -> std::optional<ObjectConfig> {
     for (const auto& conf : ObjectConfigPool) {
         if (conf.id == id) {
             return conf;
