+++ src/util/qpdf_binding.cpp
@@ -44,7 +44,7 @@ auto decompress(const std::filesystem::path& pdf_file) -> std::optional<std::fil
 
     return abs_cache_path;
 }
-auto delete_page(const sru::pdf::PdfFile& pdf_file, unsigned int page_no) -> bool {
+auto delete_page(const sru::pdf::PdfFile& pdf_file, size_t page_no) -> bool {
     ++page_no;
     const auto abs_pdf_file = std::filesystem::absolute(pdf_file.getPath().lexically_normal());
 
@@ -74,7 +74,7 @@ auto delete_page(const sru::pdf::PdfFile& pdf_file, unsigned int page_no) -> boo
         return false;
     }
 }
-auto insert_page(const sru::pdf::PdfFile& pdf_file_out, unsigned int page_no, const sru::pdf::PdfFile& pdf_file_in, unsigned int new_page_no)
+auto insert_page(const sru::pdf::PdfFile& pdf_file_out, size_t page_no, const sru::pdf::PdfFile& pdf_file_in, size_t new_page_no)
     -> bool {
     ++page_no;
     ++new_page_no;
@@ -109,7 +109,7 @@ auto insert_page(const sru::pdf::PdfFile& pdf_file_out, unsigned int page_no, co
         return false;
     }
 }
-auto change_size(const sru::pdf::PdfFile& pdf_file, int size) -> bool {
+auto change_size(const sru::pdf::PdfFile& pdf_file, size_t size) -> bool {
     auto page_count = pdf_file.getRealPageCount();
     if (size > page_count) {
         change_size(pdf_file, page_count);
